+++ src/datatypes/Csv.php
@@ -112,6 +112,10 @@ class Csv extends DataType implements DataTypeInterface
             $array = Plugin::$plugin->data->findPrimaryElement($primaryElement, $array);
         }
 
+        if ($array) {
+            $this->feedData = $array;
+        }
+
         return ['success' => true, 'data' => $array];
     }