+++ src/datatypes/GoogleSheet.php
@@ -87,6 +87,10 @@ class GoogleSheet extends DataType implements DataTypeInterface
             $array = Plugin::$plugin->data->findPrimaryElement($primaryElement, $array);
         }
 
+        if ($array) {
+            $this->feedData = $array;
+        }
+
         return ['success' => true, 'data' => $array];
     }
 }