+++ src/elements/CommerceProduct.php
@@ -395,6 +395,8 @@ class CommerceProduct extends Element
             }
         }
 
+        $parseTwig = Plugin::$plugin->service->getConfig('parseTwig', $feed['id']);
+
         foreach ($variantData as $variantContent) {
             $attributeData = [];
             $fieldData = [];
@@ -404,7 +406,7 @@ class CommerceProduct extends Element
                 if (Hash::get($fieldInfo, 'attribute')) {
                     $attributeValue = DataHelper::fetchValue(Hash::get($fieldInfo, 'data'), $fieldInfo, $this->feed);
 
-                    $attributeData[$fieldHandle] = $attributeValue;
+                    $attributeData[$fieldHandle] = $parseTwig ? DataHelper::parseFieldDataForElement($attributeValue, $this->element) : $attributeValue;
                 }
             }