+++ src/fields/Assets.php
@@ -89,7 +89,7 @@ class Assets extends Field implements FieldInterface
         $fields = Hash::get($this->fieldInfo, 'fields');
         $nativeFields = Hash::get($this->fieldInfo, 'nativeFields');
         $node = Hash::get($this->fieldInfo, 'node');
-        $nodeKey = null;
+        $nodeKey = $this->getArrayKeyFromNode($node);
 
         // Get folder id's for connecting
         $folderIds = $this->field->resolveDynamicPathToFolderId($this->element);
@@ -119,6 +119,10 @@ class Assets extends Field implements FieldInterface
         $base64ToUpload = [];
 
         $filenamesFromFeed = $upload ? DataHelper::fetchArrayValue($this->feedData, $this->fieldInfo, 'options.filenameNode') : null;
+        if ($filenamesFromFeed) {
+            // see https://github.com/craftcms/feed-me/issues/1471
+            $filenamesFromFeed = array_splice($filenamesFromFeed, $nodeKey, count($value));
+        }
 
         // Fire an 'onAssetFilename' event
         $event = new AssetFilenameEvent([
@@ -178,7 +182,14 @@ class Assets extends Field implements FieldInterface
                     $urlsToUpload[$key]['value'] = $dataValue;
 
                     if (isset($filenamesFromFeed[$key])) {
-                        $filename = $filenamesFromFeed[$key] . '.' . AssetHelper::getRemoteUrlExtension($urlsToUpload[$key]['value']);
+                        $filename = $filenamesFromFeed[$key];
+
+                        // if we can determine the extension of the remote file, use that extension
+                        $remoteUrlExtension = AssetHelper::getRemoteUrlExtension($urlsToUpload[$key]['value']);
+                        if (!empty($remoteUrlExtension)) {
+                            $filename .= '.' . $remoteUrlExtension;
+                        }
+
                         $urlsToUpload[$key]['newFilename'] = $filename;
                     } else {
                         $filename = AssetHelper::getRemoteUrlFilename($dataValue);
@@ -217,8 +228,6 @@ class Assets extends Field implements FieldInterface
                     Plugin::info('Skipping asset upload (already exists).');
                 }
             }
-
-            $nodeKey = $this->getArrayKeyFromNode($node);
         }
 
         if ($upload) {