+++ src/fields/Matrix.php
@@ -53,6 +53,10 @@ class Matrix extends Field implements FieldInterface
 
         $blocks = Hash::get($this->fieldInfo, 'blocks');
 
+        if ($blocks === null) {
+            return null;
+        }
+
         // Before we do anything, we need to extract the data from our feed and normalise it. This is especially
         // complex due to sub-fields, which each can be a variety of fields and formats, compounded by multiple or
         // Matrix blocks - we don't know! We also need to be careful of the order data is in the feed to be
@@ -97,16 +101,16 @@ class Matrix extends Field implements FieldInterface
                 } else {
                     $fieldData[$key] = $parsedValue;
                 }
-            }
 
-            foreach ($blocks as $blockHandle => $fields) {
-                foreach ($fields['fields'] as $fieldHandle => $fieldInfo) {
-                    $node = Hash::get($fieldInfo, 'node');
-                    if ($node === 'usedefault') {
-                        $key = $this->_getBlockKey($nodePathSegments, $blockHandle, $fieldHandle);
+                foreach ($blocks as $blockHandle => $fields) {
+                    foreach ($fields['fields'] as $fieldHandle => $fieldInfo) {
+                        $node = Hash::get($fieldInfo, 'node');
+                        if ($node === 'usedefault') {
+                            $key = $this->_getBlockKey($nodePathSegments, $blockHandle, $fieldHandle);
 
-                        $parsedValue = DataHelper::fetchSimpleValue($this->feedData, $fieldInfo);
-                        $fieldData[$key] = $parsedValue;
+                            $parsedValue = DataHelper::fetchSimpleValue($this->feedData, $fieldInfo);
+                            $fieldData[$key] = $parsedValue;
+                        }
                     }
                 }
             }
@@ -170,11 +174,16 @@ class Matrix extends Field implements FieldInterface
         $index = 1;
         $resultBlocks = [];
         foreach ($expanded as $blockData) {
-            // all the fields are empty and setEmptyValues is off, ignore the block
+            // if all the fields are empty and setEmptyValues is off, ignore the block
             if (
                 !empty(array_filter(
                     $blockData['fields'],
-                    fn($value) => (is_string($value) && !empty($value)) || (is_array($value) && !empty(array_filter($value)))
+                    fn($value) => (
+                        (is_string($value) && !empty($value)) ||
+                        (is_array($value) && !empty(array_filter($value))) ||
+                        is_bool($value) ||
+                        is_numeric($value)
+                    )
                 ))
             ) {
                 $resultBlocks['new' . $index++] = $blockData;