+++ src/migrations/Install.php
@@ -28,6 +28,7 @@ class Install extends Migration
 
     protected function createTables(): void
     {
+        $this->archiveTableIfExists('{{%feedme_feeds}}');
         $this->createTable('{{%feedme_feeds}}', [
             'id' => $this->primaryKey(),
             'name' => $this->string()->notNull(),
@@ -52,10 +53,25 @@ class Install extends Migration
             'dateUpdated' => $this->dateTime()->notNull(),
             'uid' => $this->uid(),
         ]);
+
+        // @see \craft\feedme\migrations\m240611_134740_create_logs_table
+        $this->archiveTableIfExists('{{%feedme_logs}}');
+        $this->createTable('{{%feedme_logs}}', [
+            'id' => $this->bigPrimaryKey(),
+            'level' => $this->integer(),
+            'category' => $this->string(),
+            'log_time' => $this->double(),
+            'prefix' => $this->text(),
+            'message' => $this->text(),
+        ]);
+
+        $this->createIndex('idx_log_level', '{{%feedme_logs}}', 'level');
+        $this->createIndex('idx_log_category', '{{%feedme_logs}}', 'category');
     }
 
     protected function removeTables(): void
     {
         $this->dropTableIfExists('{{%feedme_feeds}}');
+        $this->dropTableIfExists('{{%feedme_logs}}');
     }
 }